requirements.txt
gitdb==4.0.11
GitPython==3.1.43
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
Pygments==2.18.0
PyYAML==6.0.1
rich==13.7.1
smmap==5.0.1
textual==0.71.0
typing_extensions==4.12.2
uc-micro-py==1.0.3
